+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.7.3dev
 
+* [Fix] Fix bug that caused `pkgmt check` to fail when `pyproject.toml` had an `env_name` key
+
 ## 0.7.2 (2023-07-27)
 
 * [Feature] Added `push` and `tag` versioner configuration keys to `pyproject.toml` ([#63](https://github.com/ploomber/pkgmt/issues/63))